Keep originally empty dirs and other files from cleanup

+++ src/App/App.php
@@ -9,6 +9,14 @@ class App
 
     protected $storedFilesList = array();
 
+    /**
+     * File extensions contains a list of file extensions (with leading dot: '.php') to cleanup after store.
+     * Empty list means cleanup for all files.
+     *
+     * @var string[]
+     */
+    public $fileExtensions = [];
+
     protected function getAbsoluteFilename($filename)
     {
         $path = [];
@@ -31,7 +39,8 @@ class App
         return (DIRECTORY_SEPARATOR === '\\' ? '' : '/') . join('/', $path);
     }
 
-    protected function putContents($filepath, $content) {
+    protected function putContents($filepath, $content)
+    {
         $filepath = $this->getAbsoluteFilename($filepath);
         $this->storedFilesList[$filepath] = $filepath;
 
@@ -51,8 +60,12 @@ class App
         file_put_contents($filepath, $content);
     }
 
+    private $emptyDirs = [];
+
     public function store($path)
     {
+        $this->emptyDirs = $this->emptyDirs($path);
+
         if (DIRECTORY_SEPARATOR === '\\') {
             $path = str_replace('\\', '/', $path);
         }
@@ -64,6 +77,55 @@ class App
         }
 
         $this->clearOldFiles($path);
+    }
+
+    private function emptyDirs($path)
+    {
+        if (!file_exists($path)) {
+            return [];
+        }
+        $rii = new \RecursiveIteratorIterator(new \RecursiveDirectoryIterator($path, \FilesystemIterator::UNIX_PATHS));
+
+        $dirsToCheck = [];
+
+        /** @var \RecursiveDirectoryIterator $file */
+        foreach ($rii as $file) {
+            if ($file->isDir()) {
+                $dirsToCheck[dirname($file->getPathname())] = true;
+                continue;
+            }
+        }
+
+        $result = [];
+
+        // finding empty dirs
+        krsort($dirsToCheck);
+        foreach ($dirsToCheck as $dir => $tmp) {
+            if (!file_exists($dir)) {
+                continue;
+            }
+            $s = scandir($dir);
+            if (count($s) == 2) {
+                $result[$dir] = true;
+            }
+        }
+
+        return $result;
+    }
+
+    private function fileIsRemovable($filepath)
+    {
+        if (empty($this->fileExtensions)) {
+            return true;
+        }
+
+        foreach ($this->fileExtensions as $ext) {
+            if (substr($filepath, -strlen($ext)) === $ext) {
+                return true;
+            }
+        }
+
+        return false;
     }
 
     public function clearOldFiles($path)
@@ -83,7 +145,7 @@ class App
             }
 
             $filepath = $this->getAbsoluteFilename($file->getPathname());
-            if (!isset($this->storedFilesList[$filepath])) {
+            if (!isset($this->storedFilesList[$filepath]) && $this->fileIsRemovable($filepath)) {
                 unlink($filepath);
             }
         }
@@ -92,7 +154,7 @@ class App
         // removing empty dirs
         krsort($dirsToCheck);
         foreach ($dirsToCheck as $dir => $tmp) {
-            if (!file_exists($dir)) {
+            if (!file_exists($dir) || isset($this->emptyDirs[$dir])) {
                 continue;
             }
             $s = scandir($dir);
